Blank answer fixes + stop removing local answers on delegate
